1225- 完成地图自动加载 销毁  ai自动加载（ai行动路径未定） 任务面板（滑动暂时有缺陷）  数据更新

+++ Assets/Scripts/Models/ballSkills.cs
@@ -20,11 +20,11 @@ namespace Assets.Scripts.Models
         /// <summary>
         /// 速度（cm/s）
         /// </summary>
-        public int speed;
+        public double speed;
         /// <summary>
         /// 力量
         /// </summary>
-        public int power;
+        public double power;
         /// <summary>
         /// 能力升级花费
         /// </summary>
